Refactor product listing feature subscriber to have a more service oriented approach
