+++ src/Core/Content/Product/SalesChannel/Listing/Filter/ManufacturerFilterHandler.php
@@ -1,0 +1,60 @@
+<?php declare(strict_types=1);
+
+namespace Shopware\Core\Content\Product\SalesChannel\Listing\Filter;
+
+use Shopware\Core\Content\Product\SalesChannel\Listing\Filter;
+use Shopware\Core\Content\Product\SalesChannel\Listing\Processor\EmptyProcessTrait;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Aggregation\Metric\EntityAggregation;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsAnyFilter;
+use Shopware\Core\Framework\Log\Package;
+use Shopware\Core\Framework\Plugin\Exception\DecorationPatternException;
+use Shopware\Core\System\SalesChannel\SalesChannelContext;
+use Symfony\Component\HttpFoundation\Request;
+
+#[Package('inventory')]
+class ManufacturerFilterHandler extends AbstractFilterHandler
+{
+    use EmptyProcessTrait;
+
+    public function getDecorated(): AbstractFilterHandler
+    {
+        throw new DecorationPatternException(self::class);
+    }
+
+    public function create(Request $request, SalesChannelContext $context): ?Filter
+    {
+        if (!$request->request->get('manufacturer-filter', true)) {
+            return null;
+        }
+
+        $ids = $this->getManufacturerIds($request);
+
+        return new Filter(
+            'manufacturer',
+            !empty($ids),
+            [new EntityAggregation('manufacturer', 'product.manufacturerId', 'product_manufacturer')],
+            new EqualsAnyFilter('product.manufacturerId', $ids),
+            $ids
+        );
+    }
+
+    /**
+     * @return list<string>
+     */
+    private function getManufacturerIds(Request $request): array
+    {
+        $ids = $request->query->get('manufacturer', '');
+        if ($request->isMethod(Request::METHOD_POST)) {
+            $ids = $request->request->get('manufacturer', '');
+        }
+
+        if (\is_string($ids)) {
+            $ids = explode('|', $ids);
+        }
+
+        /** @var list<string> $ids */
+        $ids = array_filter((array) $ids);
+
+        return $ids;
+    }
+}
