+++ src/Core/Content/Product/SalesChannel/Listing/Filter/PriceFilterHandler.php
@@ -1,0 +1,52 @@
+<?php declare(strict_types=1);
+
+namespace Shopware\Core\Content\Product\SalesChannel\Listing\Filter;
+
+use Shopware\Core\Content\Product\SalesChannel\Listing\Filter;
+use Shopware\Core\Content\Product\SalesChannel\Listing\Processor\EmptyProcessTrait;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Aggregation\Metric\StatsAggregation;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\RangeFilter;
+use Shopware\Core\Framework\Log\Package;
+use Shopware\Core\Framework\Plugin\Exception\DecorationPatternException;
+use Shopware\Core\System\SalesChannel\SalesChannelContext;
+use Symfony\Component\HttpFoundation\Request;
+
+#[Package('inventory')]
+class PriceFilterHandler extends AbstractFilterHandler
+{
+    use EmptyProcessTrait;
+
+    public function getDecorated(): AbstractFilterHandler
+    {
+        throw new DecorationPatternException(self::class);
+    }
+
+    public function create(Request $request, SalesChannelContext $context): ?Filter
+    {
+        if (!$request->request->get('price-filter', true)) {
+            return null;
+        }
+
+        $min = $request->get('min-price');
+        $max = $request->get('max-price');
+
+        $range = [];
+        if ($min !== null && $min >= 0) {
+            $range[RangeFilter::GTE] = $min;
+        }
+        if ($max !== null && $max >= 0) {
+            $range[RangeFilter::LTE] = $max;
+        }
+
+        return new Filter(
+            'price',
+            !empty($range),
+            [new StatsAggregation('price', 'product.cheapestPrice', true, true, false, false)],
+            new RangeFilter('product.cheapestPrice', $range),
+            [
+                'min' => (float) $request->get('min-price'),
+                'max' => (float) $request->get('max-price'),
+            ]
+        );
+    }
+}
