+++ src/Core/Content/Product/SalesChannel/Listing/Filter/RatingFilterHandler.php
@@ -1,0 +1,49 @@
+<?php declare(strict_types=1);
+
+namespace Shopware\Core\Content\Product\SalesChannel\Listing\Filter;
+
+use Shopware\Core\Content\Product\SalesChannel\Listing\Filter;
+use Shopware\Core\Content\Product\SalesChannel\Listing\Processor\EmptyProcessTrait;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Aggregation\Bucket\FilterAggregation;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Aggregation\Metric\MaxAggregation;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\RangeFilter;
+use Shopware\Core\Framework\Log\Package;
+use Shopware\Core\Framework\Plugin\Exception\DecorationPatternException;
+use Shopware\Core\System\SalesChannel\SalesChannelContext;
+use Symfony\Component\HttpFoundation\Request;
+
+#[Package('inventory')]
+class RatingFilterHandler extends AbstractFilterHandler
+{
+    use EmptyProcessTrait;
+
+    public function getDecorated(): AbstractFilterHandler
+    {
+        throw new DecorationPatternException(self::class);
+    }
+
+    public function create(Request $request, SalesChannelContext $context): ?Filter
+    {
+        if (!$request->request->get('rating-filter', true)) {
+            return null;
+        }
+
+        $filtered = $request->get('rating');
+
+        return new Filter(
+            'rating',
+            $filtered !== null,
+            [
+                new FilterAggregation(
+                    'rating-exists',
+                    new MaxAggregation('rating', 'product.ratingAverage'),
+                    [new RangeFilter('product.ratingAverage', [RangeFilter::GTE => 0])]
+                ),
+            ],
+            new RangeFilter('product.ratingAverage', [
+                RangeFilter::GTE => (int) $filtered,
+            ]),
+            $filtered
+        );
+    }
+}
