+++ src/Core/Content/Product/SalesChannel/Listing/Filter/ShippingFreeFilterHandler.php
@@ -1,0 +1,47 @@
+<?php declare(strict_types=1);
+
+namespace Shopware\Core\Content\Product\SalesChannel\Listing\Filter;
+
+use Shopware\Core\Content\Product\SalesChannel\Listing\Filter;
+use Shopware\Core\Content\Product\SalesChannel\Listing\Processor\EmptyProcessTrait;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Aggregation\Bucket\FilterAggregation;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Aggregation\Metric\MaxAggregation;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
+use Shopware\Core\Framework\Log\Package;
+use Shopware\Core\Framework\Plugin\Exception\DecorationPatternException;
+use Shopware\Core\System\SalesChannel\SalesChannelContext;
+use Symfony\Component\HttpFoundation\Request;
+
+#[Package('inventory')]
+class ShippingFreeFilterHandler extends AbstractFilterHandler
+{
+    use EmptyProcessTrait;
+
+    public function getDecorated(): AbstractFilterHandler
+    {
+        throw new DecorationPatternException(self::class);
+    }
+
+    public function create(Request $request, SalesChannelContext $context): ?Filter
+    {
+        if (!$request->request->get('shipping-free-filter', true)) {
+            return null;
+        }
+
+        $filtered = (bool) $request->get('shipping-free', false);
+
+        return new Filter(
+            'shipping-free',
+            $filtered === true,
+            [
+                new FilterAggregation(
+                    'shipping-free-filter',
+                    new MaxAggregation('shipping-free', 'product.shippingFree'),
+                    [new EqualsFilter('product.shippingFree', true)]
+                ),
+            ],
+            new EqualsFilter('product.shippingFree', true),
+            $filtered
+        );
+    }
+}
