+++ src/Core/Content/Product/SalesChannel/Listing/Processor/AssociationLoadingProcessor.php
@@ -1,0 +1,26 @@
+<?php declare(strict_types=1);
+
+namespace Shopware\Core\Content\Product\SalesChannel\Listing\Processor;
+
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
+use Shopware\Core\Framework\Log\Package;
+use Shopware\Core\Framework\Plugin\Exception\DecorationPatternException;
+use Shopware\Core\System\SalesChannel\SalesChannelContext;
+use Symfony\Component\HttpFoundation\Request;
+
+#[Package('inventory')]
+class AssociationLoadingProcessor extends AbstractListingProcessor
+{
+    use EmptyProcessTrait;
+
+    public function getDecorated(): AbstractListingProcessor
+    {
+        throw new DecorationPatternException(self::class);
+    }
+
+    public function prepare(Request $request, Criteria $criteria, SalesChannelContext $context): void
+    {
+        $criteria->addAssociation('manufacturer');
+        $criteria->addAssociation('options');
+    }
+}
