+++ src/Core/Content/Product/SalesChannel/Listing/Processor/EmptyProcessTrait.php
@@ -1,0 +1,16 @@
+<?php declare(strict_types=1);
+
+namespace Shopware\Core\Content\Product\SalesChannel\Listing\Processor;
+
+use Shopware\Core\Content\Product\SalesChannel\Listing\ProductListingResult;
+use Shopware\Core\Framework\Log\Package;
+use Shopware\Core\System\SalesChannel\SalesChannelContext;
+use Symfony\Component\HttpFoundation\Request;
+
+#[Package('inventory')]
+trait EmptyProcessTrait
+{
+    public function process(Request $request, ProductListingResult $result, SalesChannelContext $context): void
+    {
+    }
+}
