+++ src/Core/Content/Product/SalesChannel/Listing/Processor/HeaderProcessor.php
@@ -1,0 +1,39 @@
+<?php declare(strict_types=1);
+
+namespace Shopware\Core\Content\Product\SalesChannel\Listing\Processor;
+
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
+use Shopware\Core\Framework\Log\Package;
+use Shopware\Core\Framework\Plugin\Exception\DecorationPatternException;
+use Shopware\Core\System\SalesChannel\SalesChannelContext;
+use Symfony\Component\HttpFoundation\Request;
+
+#[Package('inventory')]
+class HeaderProcessor extends AbstractListingProcessor
+{
+    use EmptyProcessTrait;
+
+    public function getDecorated(): AbstractListingProcessor
+    {
+        throw new DecorationPatternException(self::class);
+    }
+
+    public function prepare(Request $request, Criteria $criteria, SalesChannelContext $context): void
+    {
+        if ($request->get('no-aggregations')) {
+            $criteria->resetAggregations();
+        }
+
+        if ($request->get('only-aggregations')) {
+            // set limit to zero to fetch no products.
+            $criteria->setLimit(0);
+
+            // no total count required
+            $criteria->setTotalCountMode(Criteria::TOTAL_COUNT_MODE_NONE);
+
+            // sorting and association are only required for the product data
+            $criteria->resetSorting();
+            $criteria->resetAssociations();
+        }
+    }
+}
