+++ src/Core/Content/Product/SalesChannel/Listing/Processor/PagingProcessor.php
@@ -1,0 +1,69 @@
+<?php declare(strict_types=1);
+
+namespace Shopware\Core\Content\Product\SalesChannel\Listing\Processor;
+
+use Shopware\Core\Content\Product\SalesChannel\Listing\ProductListingResult;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
+use Shopware\Core\Framework\Log\Package;
+use Shopware\Core\Framework\Plugin\Exception\DecorationPatternException;
+use Shopware\Core\System\SalesChannel\SalesChannelContext;
+use Shopware\Core\System\SystemConfig\SystemConfigService;
+use Symfony\Component\HttpFoundation\Request;
+
+#[Package('inventory')]
+class PagingProcessor extends AbstractListingProcessor
+{
+    /**
+     * @internal
+     */
+    public function __construct(private readonly SystemConfigService $config)
+    {
+    }
+
+    public function getDecorated(): AbstractListingProcessor
+    {
+        throw new DecorationPatternException(self::class);
+    }
+
+    public function prepare(Request $request, Criteria $criteria, SalesChannelContext $context): void
+    {
+        $limit = $this->getLimit($request, $context);
+
+        $page = $this->getPage($request);
+
+        $criteria->setOffset(($page - 1) * $limit);
+        $criteria->setLimit($limit);
+        $criteria->setTotalCountMode(Criteria::TOTAL_COUNT_MODE_EXACT);
+    }
+
+    public function process(Request $request, ProductListingResult $result, SalesChannelContext $context): void
+    {
+        $result->setPage($this->getPage($request));
+
+        $result->setLimit($this->getLimit($request, $context));
+    }
+
+    private function getLimit(Request $request, SalesChannelContext $context): int
+    {
+        $limit = $request->query->getInt('limit');
+
+        if ($request->isMethod(Request::METHOD_POST)) {
+            $limit = $request->request->getInt('limit', $limit);
+        }
+
+        $limit = $limit > 0 ? $limit : $this->config->getInt('core.listing.productsPerPage', $context->getSalesChannel()->getId());
+
+        return $limit <= 0 ? 24 : $limit;
+    }
+
+    private function getPage(Request $request): int
+    {
+        $page = $request->query->getInt('p', 1);
+
+        if ($request->isMethod(Request::METHOD_POST)) {
+            $page = $request->request->getInt('p', $page);
+        }
+
+        return $page <= 0 ? 1 : $page;
+    }
+}
