+++ src/Core/Content/Product/SalesChannel/Listing/ProductListingRoute.php
@@ -12,6 +12,7 @@ use Shopware\Core\Framework\DataAbstractionLayer\EntityRepository;
 use Shopware\Core\Framework\DataAbstractionLayer\PartialEntity;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
+use Shopware\Core\Framework\Feature;
 use Shopware\Core\Framework\Log\Package;
 use Shopware\Core\Framework\Plugin\Exception\DecorationPatternException;
 use Shopware\Core\System\SalesChannel\SalesChannelContext;
@@ -67,9 +68,11 @@ class ProductListingRoute extends AbstractProductListingRoute
 
         $result->addCurrentFilter('navigationId', $categoryId);
 
-        $this->eventDispatcher->dispatch(
-            new ProductListingResultEvent($request, $result, $context)
-        );
+        if (!Feature::isActive('v6.6.0.0')) {
+            $this->eventDispatcher->dispatch(
+                new ProductListingResultEvent($request, $result, $context)
+            );
+        }
 
         $result->setStreamId($category->get('productStreamId'));
